rename N folders
