+++ src/components/loader/loader.component.tsx
