+++ src/components/loader/loader.styles.ts
