+++ src/pages/paint/components/buttons-panel/buttons-panel.component.tsx
