+++ src/pages/paint/components/buttons-panel/buttons-panel.styles.ts
