+++ src/pages/paint/components/buttons-panel/index.ts
