+++ src/pages/paint/components/drawing-tools-panel/drawing-tools-panel.component.tsx
