+++ src/pages/paint/components/drawing-tools-panel/index.ts
