+++ src/pages/paint/index.ts
