+++ src/pages/paint/paint.component.tsx
