+++ src/pages/paint/paint.styles.ts
